Fixes stacking when pressing Enter
Stacking dialogs had an additionnal side effect: when you cancel the last one, you can't click anywhere else in app (probably the layer not properly cleaned up). So as this commit will fix an actual issue (you don't want to re-submit the background button/link when pressing enter in a dialog), it may not be fixing the whole thing.